Rearrange code to prevent circular import
