+++ helios/horses/__init__.py
@@ -1,3 +1,4 @@
 # from .race import Race, RaceHorse
-from .horse import Horse
 from .enumerations import BetType
+from .horse import Horse
+from .race import RaceHorse, Race
